Changed walls
Made walls meshes with wireframe materials that are invisible in game. The goal of this is to allow bullets to pass through them, but not vehicles. Bullets hit and stick to them right now. Cannot figure out how to change max speeds of AI, despite changing "Maxspeed" variable in parent in instances blueprints.